requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.109.2
google-api-python-client==2.123.0
google-auth-oauthlib==1.2.0
grequests==0.7.0
pandas==1.5.3
Pillow==9.3.0
plotly==5.19.0
requests==2.27.1
st-pages==0.4.5
streamlit==1.32.0
streamlit-antd-components==0.3.2
streamlit-autorefresh==1.0.1
streamlit-extras==0.4.0
